resolving TODOs from review
- removed pathSpec from Request
- getServletMapping moved to ServletHandler
